fix: improve circularity detection
Improve circularity detection in the jwz algorithm by also considering
the siblings of the children. The current implementation will only
follow a direct line of children but does not consider the siblings.

With this change, the threading algorithm also picks up all of the 2405
messages in the test set (counting the messages with: ls -1
test/testdata/ham | wc -l). Before, the number was hardcoded to 2387
(and there was no hint as to why the hardcoded number was lower than the
total number of test messages). The examples and tests have been
adjusted accordingly.

Signed-off-by: Koni Marti <koni.marti@gmail.com>